Inline AbstractServerImplBuilder: the rest and properties
